moved queue and tryEncode to new files
